Refactor M4 Prover/Verifier to delegate to IOPProver/IOPVerifier
The M4 Prover/Verifier structs previously bundled the constraint-system
state and the BaseFold compiler together, and their prove()/verify()
methods inlined the trace commit, the reduction, and the ring-switch
opening directly, calling out to free-standing prove_reduction /
verify_reduction functions for the reduction step. This diverged from the
binius-prover / binius-verifier crates, which split each top-level struct
into a channel-abstraction-agnostic IOPProver/IOPVerifier (holding just
the constraint-system-derived state, with a Channel-generic prove/verify
method) and a thin outer wrapper that only creates the channel from the
transcript, delegates, and finishes it.

Introduce binius_m4_prover::IOPProver and binius_m4_verifier::IOPVerifier
following that same split:

- IOPVerifier { cs, layout }: its verify() folds together the former
  verify_reduction body (AND-check, IntMul check, instance
  re-randomization, shift reduction, public-input check) with the
  trace-oracle receive and the ring-switch verification/oracle-relation
  opening that used to live directly in Verifier::verify.
- IOPProver { cs, key_collection }: its prove() folds together the former
  prove_reduction body with the trace packing/commit and the ring-switch
  proving/oracle-relation queuing that used to live directly in
  Prover::prove.
- Verifier and Prover<P> are now thin wrappers: they hold the IOP
  struct plus the BaseFold compiler, and their prove()/verify() methods
  only build the channel from the transcript, delegate to
  iop_prover.prove() / iop_verifier.verify(), and finish the channel.
- IOPProver::new(iop_verifier, key_collection) mirrors binius-prover's
  constructor, consuming a cloned IOPVerifier; IOPVerifier gained a
  Clone derive and an into_constraint_system() accessor to support this.

Both crates' reduction.rs files are deleted; the private re-randomization
helper types (Operation/OperationClaim, RerandomizedOperations) moved
into prove.rs/verify.rs since prove_reduction/verify_reduction no longer
exist as a separate public API.

reduction.rs's round-trip tests (MUL constraints, non-power-of-two
constant counts, mixed AND/MUL constraint counts with wide packing,
transcript tampering) exercised the granular prove_reduction/
verify_reduction entry points directly over Naive channels, asserting on
internal values like r_rho and the reduced witness evaluation. Since
those entry points are gone, the tests are ported to full Prover/Verifier
protocol round trips in prove.rs, checking round-trip success/failure
rather than internal reduction state — mirroring how
crates/prover/tests/prove_verify.rs only tests the analogous
IntMul/BitAnd/shift composition at the full Prover/Verifier level.

Verified with cargo build --workspace, cargo test -p binius-m4-prover -p
binius-m4-verifier (35 tests passing), cargo clippy -D warnings, cargo
doc --document-private-items, and cargo +nightly-2026-01-01 fmt.
