+++ crates/m4-prover/src/lib.rs
@@ -5,13 +5,11 @@
 
 mod operand_witness;
 mod prove;
-mod reduction;
 mod shift;
 #[cfg(test)]
 mod test_utils;
 mod value_table;
 
 pub use operand_witness::BatchAndCheckWitness;
-pub use prove::Prover;
-pub use reduction::{ReductionProverOutput, prove_reduction};
+pub use prove::{IOPProver, Prover};
 pub use value_table::{BatchWitnessFiller, ValueTable};
