+++ crates/m4-verifier/src/lib.rs
@@ -4,10 +4,8 @@
 
 mod bitand;
 mod commit;
-mod reduction;
 mod verify;
 
 pub use bitand::verify_bitand_reduction;
 pub use commit::BatchCommitLayout;
-pub use reduction::{ReductionVerifierOutput, verify_reduction};
-pub use verify::Verifier;
+pub use verify::{IOPVerifier, Verifier};
